Refactoring in admin/configuration. Fix issue preventing guest user configuration to be saved.

+++ src/LegacyPages/install.php
@@ -394,7 +394,9 @@ if ($step === 'user') {
             rename(PHPWG_ROOT_PATH . 'local/config/database.inc.tmp.php', PHPWG_ROOT_PATH . 'local/config/database.inc.php');
 
             // remove temporary upgrade file needed for upgrade from previous releases
-            unlink(PHPWG_ROOT_PATH . 'upgrade.php');
+            if (is_writable(PHPWG_ROOT_PATH . 'upgrade.php')) {
+                unlink(PHPWG_ROOT_PATH . 'upgrade.php');
+            }
             $infos[] = \Phyxo\Functions\Language::l10n('Congratulations, Phyxo installation is completed');
         } catch (\Exception $e) {
             $errors[] = \Phyxo\Functions\Language::l10n($e->getMessage());
